+++ app/src/main/java/ru/noties/tumbleweed/tutorial/scene/AnimationScene_Text2.java
@@ -63,7 +63,7 @@ public class AnimationScene_Text2 extends AnimationScene {
 
         @Override
         public void getValues(@NonNull TextView textView, @NonNull float[] values) {
-            values[0] = Float.parseFloat(textView.getText().toString());
+            values[0] = Integer.parseInt(textView.getText().toString());
         }
 
         @Override
